publicui-core/src/test/java/com/funnelback/publicui/search/model/collection/facetednavigation/impl/GScopeItemTest.java
package com.funnelback.publicui.search.model.collection.facetednavigation.impl;

import java.util.Collections;
import java.util.List;

import org.junit.Assert;
import org.junit.Before;
import org.junit.Test;

import com.funnelback.publicui.search.model.collection.QueryProcessorOption;
import com.google.common.collect.Sets;

public class GScopeItemTest {

    private GScopeItem category;
    
    @Before
    public void before() {
        category = new GScopeItem("categoryName", 12);
        category.setFacetName("facetName");
        category.setLabel("label");
    }
    
    @Test
    public void testFields() {
        Assert.assertEquals("categoryName", category.getData());
        Assert.assertEquals(12, category.getGScopeNumber());
        Assert.assertEquals(12, category.getUserSetGScope());
        Assert.assertEquals("facetName", category.getFacetName());
        Assert.assertEquals("label", category.getLabel());
        
        Assert.assertEquals(Sets.newHashSet("f.facetName|12"), category.getAllQueryStringParamNames());
    }
    
    @Test
    public void testConstraint() {
        Assert.assertEquals("12", category.getGScope1Constraint());
    }
    
    @Test
    public void testQPOs() {
        List<QueryProcessorOption<?>> actual = category.getQueryProcessorOptions(null);
        Assert.assertEquals(Collections.singletonList(new QueryProcessorOption<>("countgbits", "all")), actual);
    }
    
    
}
